backend/requirements_production.txt
# Production requirements for Browser Agent API
# Core dependencies
google-generativeai>=0.8.0
playwright>=1.51.0
flask>=3.0.3
pydantic>=2.11.0
python-dotenv>=1.0.1
aiohttp>=3.10.5
Flask-Cors>=5.0.0

# Production WSGI servers
waitress>=2.1.2
gunicorn>=21.2.0

# Additional utilities
requests>=2.32.3

# Note: After installing, run: playwright install chromium

